SRC/requirements.txt
Flask==1.1.1
pymongo==3.10.1
nltk==3.4.5
pandas==1.0.1
requests==2.21.0
jsonschema==3.2.0
python-dotenv==0.11.0
sklearn==0.0
dnspython